requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cssselect==1.1.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
django-tinymce==3.0.2
fake-useragent==0.1.11
filelock==3.0.12
FWapi==0.1
gunicorn==20.0.4
idna==2.10
kombu==4.6.11
lxml==4.5.2
parse==1.18.0
Pillow==7.2.0
pipenv==2020.8.13
psycopg2==2.8.6
pycparser==2.20
pyee==7.0.3
pyfilmweb==0.1.1.2
pyppeteer==0.2.2
pyquery==1.4.1
pytz==2020.1
rabbitmq==0.2.0
requests==2.24.0
requests-html==0.10.0
selenium==3.141.0
six==1.15.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
tqdm==4.48.2
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
w3lib==1.22.0
websockets==8.1
whitenoise==5.2.0
